garbage-calendar/requirements.txt
caldav==0.8.2
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
vobject==0.9.6.1
